WIP home

+++ src/Entity/RedPizza.php
@@ -61,7 +61,7 @@ class RedPizza
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
